# swagger.yaml
...
openapi: "3.0.3"
info:
  version: development
  title: Ticketing Application
  description: "API documentation for **Ticketing Application**"
servers:
  - url:  http://localhost:8080
    description: "Development"
...
paths:
...
  /user/update:
    post:
      tags:
        - user
      summary: update user
      operationId: updateUser
      requestBody:
        description: Request body to update user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRq'
            examples:
              ScenarioA:
                summary: "Scenario A"
                value:
                  id: 68
                  name:  "budi"
                  email: "buudi@mail.com"
        required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateUserRs'
              examples:
                ScenarioA:
                  summary: "Scenario OK"
                  value:
                    {
                        "meta": {
                            "message": "Update user data",
                            "code": 200,
                            "status": "success"
                        },
                        "data": {
                            "id": 68,
                            "name": "budi",
                            "email": "buudi@mail.com"
                        }
                    }                    
  /user/delete:
    post:
      tags:
        - user
      summary: delete user
      operationId: deleteUser
      requestBody:
        description: Request body to delete user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteUserRq'
            examples:
              ScenarioA:
                summary: "Scenario A"
                value:
                  id:  1
        required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteUserRs'
              examples:
                ScenarioA:
                  summary: "Scenario OK"
                  value:
                      {
                          "meta": {
                              "message": "Delete user data",
                              "code": 200,
                              "status": "success"
                          },
                          "data": null
                      }                                      
...
  /user:
    post:
      tags:
        - user
      summary: get user by id
      operationId: getUserById
      requestBody:
        description: Request body to get user by id
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetUserRq'
            examples:
              ScenarioA:
                summary: "Scenario A"
                value:
                  id:  66
        required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserRs'
              examples:
                ScenarioA:
                  summary: "Scenario OK"
                  value:
                    meta:
                      message: Get user data
                      code: 200
                      status: success
                    data:
                      id: 66
                      name: anton
                      email: anton@mail.com
...
components:
  schemas:
...
    UpdateUserRq:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          pattern: user@mail.com
    UpdateUserRs:
        allOf: 
          - $ref: "#/components/schemas/ApiResponse"
          - type: "object"
            properties:
              data:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  email:
                    type: string
                    pattern: user@mail.com
    DeleteUserRq:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
    DeleteUserRs:
        allOf: 
          - $ref: "#/components/schemas/ApiResponse"
          - type: "object"
            properties:
              data:
                type: object
...
    GetUserRq:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
    GetUserRs:
        allOf: 
          - $ref: "#/components/schemas/ApiResponse"
          - type: "object"
            properties:
              data:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  email:
                    type: string
                    pattern: user@mail.com                      
...
    ApiResponse:
      type: object
      properties:
        meta:
          type: object
          properties:
            message:
              type: string
            code:
              type: integer
            status:
              type: string